# src/main/java/mss/domain/entity/Suggestions.java
package mss.domain.entity;

import java.util.List;

public class Suggestions {

    private List<String> suggestions;

    public List<String> getSuggestions() {
        return suggestions;
    }

    public void setSuggestions(List<String> suggestions) {
        this.suggestions = suggestions;
    }
}